In an attempt to address https://github.com/polarsignals/frostdb/issues/246 I've rewritten the tx list to be a [Harris style ordered linked-list](https://timharris.uk/papers/2001-disc.pdf). I originally wrote the list using pointer arithmetic (shame on me) but Go's race detector screamed at me, so I changed it to use `nil` pointers and save off the original pointer in a separate atomic.Pointer.

Ordering the linked list means we no longer need to perform a periodic sweep because the ordering of the completed transactions means we only need to iterate until we encounter a tx that isn't next in line. Furthermore this fixed a race condition that was present with the deletion routine that could cause transactions to go missing which would cause the list to be permanently hung.

```
thor@thors-MacBook-Pro frostdb % go test -v -run TX -bench TX
=== RUN   Test_TXList_Mark
--- PASS: Test_TXList_Mark (0.00s)
=== RUN   Test_TXList_Basic
--- PASS: Test_TXList_Basic (0.00s)
=== RUN   Test_TXList_Async
--- PASS: Test_TXList_Async (0.00s)
goos: darwin
goarch: arm64
pkg: github.com/polarsignals/frostdb
Benchmark_TXList_InsertAndDrain
Benchmark_TXList_InsertAndDrain/10:1000
Benchmark_TXList_InsertAndDrain/10:1000-10         	      82	  14727339 ns/op
Benchmark_TXList_InsertAndDrain/100:100
Benchmark_TXList_InsertAndDrain/100:100-10         	     100	  14858850 ns/op
Benchmark_TXList_InsertAndDrain/1000:100
Benchmark_TXList_InsertAndDrain/1000:100-10        	       2	 703842230 ns/op
Benchmark_TXList_InsertAndDrain/10:100
Benchmark_TXList_InsertAndDrain/10:100-10          	    2536	    471606 ns/op
PASS
ok  	github.com/polarsignals/frostdb	7.234s
```