@thorfour 2022-12-01
```
thor@thors-MacBook-Pro frostdb % benchstat before.txt after.txt
name                                          old time/op    new time/op    delta
_Table_Insert_1000Rows_10Iters_10Writers-10      321ms ±30%      17ms ± 2%  -94.61%  (p=0.000 n=10+10)
_Table_Insert_100Rows_1000Iters_1Writers-10      195ms ± 4%     195ms ± 3%     ~     (p=0.829 n=10+8)
_Table_Insert_100Rows_100Iters_100Writers-10     37.6s ± 4%      0.4s ±15%  -98.92%  (p=0.000 n=10+10)

name                                          old alloc/op   new alloc/op   delta
_Table_Insert_1000Rows_10Iters_10Writers-10     98.1MB ± 3%    48.4MB ± 5%  -50.72%  (p=0.000 n=10+10)
_Table_Insert_100Rows_1000Iters_1Writers-10      279MB ± 9%     290MB ±11%     ~     (p=0.280 n=10+10)
_Table_Insert_100Rows_100Iters_100Writers-10    3.22GB ± 6%    2.10GB ±10%  -34.58%  (p=0.000 n=10+10)

name                                          old allocs/op  new allocs/op  delta
_Table_Insert_1000Rows_10Iters_10Writers-10       443k ± 1%      340k ± 2%  -23.34%  (p=0.000 n=10+10)
_Table_Insert_100Rows_1000Iters_1Writers-10      1.89M ± 5%     1.93M ± 6%     ~     (p=0.280 n=10+10)
_Table_Insert_100Rows_100Iters_100Writers-10     20.9M ± 2%     15.6M ± 6%  -25.39%  (p=0.000 n=10+10)
```